test: split proposals utils and services tests
