+++ frontend/svelte/src/tests/lib/utils/proposals.utils.spec.ts
@@ -1,8 +1,4 @@
 import { GovernanceCanister, ProposalInfo } from "@dfinity/nns";
-import {
-  listNextProposals,
-  listProposals,
-} from "../../../lib/services/proposals.services";
 import { proposalsStore } from "../../../lib/stores/proposals.store";
 import {
   emptyProposals,
@@ -55,67 +51,5 @@ describe("proposals-utils", () => {
     );
 
     proposalsStore.setProposals([]);
-  });
-
-  it("should call the canister to list proposals", async () => {
-    await listProposals({});
-
-    expect(spyListProposals).toHaveReturnedTimes(1);
-  });
-
-  it("should call the canister to list the next proposals", async () => {
-    await listNextProposals({
-      beforeProposal: mockProposals[mockProposals.length - 1].id,
-    });
-
-    expect(spyListProposals).toHaveReturnedTimes(1);
-  });
-
-  it("should clear the list proposals before query", async () => {
-    const spy = jest.spyOn(proposalsStore, "setProposals");
-    await listProposals({ clearBeforeQuery: true });
-    expect(spy).toHaveBeenCalledTimes(2);
-    spy.mockClear();
-  });
-
-  it("should not clear the list proposals before query", async () => {
-    const spy = jest.spyOn(proposalsStore, "setProposals");
-    await listProposals({ clearBeforeQuery: false });
-    expect(spy).toHaveBeenCalledTimes(1);
-    spy.mockClear();
-  });
-
-  it("should push new proposals to the list", async () => {
-    const spy = jest.spyOn(proposalsStore, "pushProposals");
-    await listNextProposals({
-      beforeProposal: mockProposals[mockProposals.length - 1].id,
-    });
-    expect(spy).toHaveBeenCalledTimes(1);
-    spy.mockClear();
-  });
-
-  it("should not push empty proposals to the list", async () => {
-    const mockEmptyGovernanceCanister: MockGovernanceCanister =
-      new MockGovernanceCanister([]);
-
-    jest
-      .spyOn(GovernanceCanister, "create")
-      .mockImplementation(
-        (): GovernanceCanister => mockEmptyGovernanceCanister
-      );
-
-    const spyListProposals = jest.spyOn(
-      mockEmptyGovernanceCanister,
-      "listProposals"
-    );
-
-    const spy = jest.spyOn(proposalsStore, "pushProposals");
-    await listNextProposals({
-      beforeProposal: mockProposals[mockProposals.length - 1].id,
-    });
-    expect(spy).toHaveBeenCalledTimes(0);
-    spy.mockClear();
-
-    spyListProposals.mockClear();
   });
 });
